+++ controllers/bannerlogo.js
@@ -6,9 +6,9 @@ import { deleteFile } from "../middleware/deleteFile.js";
 export const create = async (req, res, next) => {
   try {
     const mainImgaeLink = req.files.mainImgaeLink
-    ? req.files.mainImgaeLink[0].filename
-    : "";
-    const newBanner = new BannerLogo({mainImgaeLink:mainImgaeLink});
+      ? req.files.mainImgaeLink[0].filename
+      : "";
+    const newBanner = new BannerLogo({ mainImgaeLink: mainImgaeLink });
     const savedBanner = await newBanner.save();
     return res.status(200).json({ result: savedBanner }).end();
   } catch (error) {
@@ -49,19 +49,28 @@ export const editById = async (req, res, next) => {
 
     if (req.files.mainImgaeLink) {
       obj.mainImgaeLink = req.files.mainImgaeLink[0].filename;
-      const filename = existingBanner.mainImgaeLink;
-      const filePath = `/mainImage/${filename}`;
-      try {
-        await deleteFile(filePath);
-      } catch (error) {
-        return res
-          .status(400)
-          .json({ message: error.message || "Internal server error!" })
-          .end();
+      if (
+        existingBanner?.mainImgaeLink &&
+        existingBanner?.mainImgaeLink?.length > 0
+      ) {
+        const filename = existingBanner.mainImgaeLink;
+        const filePath = `/mainImage/${filename}`;
+        try {
+          await deleteFile(filePath);
+        } catch (error) {
+          return res
+            .status(400)
+            .json({ message: error.message || "Internal server error!" })
+            .end();
+        }
       }
     }
 
-    await BannerLogo.findByIdAndUpdate(req.body._id, { $set: { ...obj } },{new:true});
+    await BannerLogo.findByIdAndUpdate(
+      req.body._id,
+      { $set: { ...obj } },
+      { new: true }
+    );
 
     return res.status(200).json({ message: "Successfully Updated" }).end();
   } catch (error) {
@@ -86,12 +95,15 @@ export const deleteById = async (req, res, next) => {
 
     await BannerLogo.findByIdAndDelete(req.params.id);
 
-    if(existingBannerLogo?.mainImgaeLink){
+    if (
+      existingBannerLogo?.mainImgaeLink &&
+      existingBannerLogo?.mainImgaeLink?.length > 0
+    ) {
       const filename = existingBannerLogo.mainImgaeLink;
       const filePath = `/mainImage/${filename}`;
       try {
         const response = await deleteFile(filePath);
-        console.log(response)
+        console.log(response);
       } catch (error) {
         return res
           .status(400)
