+++ controllers/city.js
@@ -4,12 +4,13 @@ import CityModel from "../model/City.js";
 import PropertyModel from "../model/Property.js";
 import mongoose from "mongoose";
 import { deleteFile } from "../middleware/deleteFile.js";
+import Property from "../model/Property.js";
 
 export const create = async (req, res, next) => {
   try {
     const mainImgaeLink = req.files.mainImgaeLink
       ? req.files.mainImgaeLink[0].filename
-      : ""; 
+      : "";
     const newCity = new CityModel({
       ...req.body,
       mainImgaeLink: mainImgaeLink,
@@ -61,17 +62,22 @@ export const editById = async (req, res, next) => {
     let obj = {
       ...req.body,
     };
-    if (req.files.mainImgaeLink && existingCity.mainImgaeLink.length > 0) {
+    if (req.files.mainImgaeLink) {
       obj.mainImgaeLink = req.files.mainImgaeLink[0].filename;
-      const filename = existingCity.mainImgaeLink;
-      const filePath = `/mainImage/${filename}`;
-      try {
-        await deleteFile(filePath);
-      } catch (error) {
-        return res
-          .status(400)
-          .json({ message: error.message || "Internal server error!" })
-          .end();
+      if (
+        existingCity?.mainImgaeLink &&
+        existingCity?.mainImgaeLink?.length > 0
+      ) {
+        const filename = existingCity.mainImgaeLink;
+        const filePath = `/mainImage/${filename}`;
+        try {
+          await deleteFile(filePath);
+        } catch (error) {
+          return res
+            .status(400)
+            .json({ message: error.message || "Internal server error!" })
+            .end();
+        }
       }
     }
 
@@ -102,14 +108,23 @@ export const deleteById = async (req, res, next) => {
       return res.status(400).json({ message: "City Not Exist!!" }).end();
     }
 
+    const isExistingCity = await Property.findOne({cityRef:new mongoose.Types.ObjectId(req.params.id)});
+
+    if(isExistingCity){
+      return res.status(400).json({ message: "Already property avalible under the city" }).end();
+    }
+
     await CityModel.findByIdAndDelete(req.params.id);
 
-    if(existingCity?.mainImgaeLink){
+    if (
+      existingCity?.mainImgaeLink &&
+      existingCity?.mainImgaeLink?.length > 0
+    ) {
       const filename = existingCity.mainImgaeLink;
       const filePath = `/mainImage/${filename}`;
       try {
         const response = await deleteFile(filePath);
-        console.log(response)
+        console.log(response);
       } catch (error) {
         return res
           .status(400)
@@ -117,7 +132,6 @@ export const deleteById = async (req, res, next) => {
           .end();
       }
     }
-
 
     return res.status(200).json({ message: "Successfully Deleted" }).end();
   } catch (error) {
